Se realiza endpoint de login + service y mapper para login

+++ src/main/java/com/backProyectoFinal/Entity/Dto/usuario/UsuarioDto.java
@@ -18,7 +18,7 @@ public class UsuarioDto {
     private String nombre;
     private String apellido;
     private String email;
-    private int celular;
+    private Long celular;
     private String contrasenia;
     private Rol rol;
 }
